Tolerate BundleException
