+++ dev/com.ibm.ws.wsat.concurrent_fat/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2019, 2021 IBM Corporation and others.
+# Copyright (c) 2019, 2022 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -33,6 +33,7 @@ tested.features: \
   com.ibm.websphere.org.osgi.core,\
   com.ibm.websphere.org.osgi.service.component,\
   com.ibm.ws.transaction.test.util;version=latest,\
+  com.ibm.ws.transaction.fat.util;version=latest,\
   com.ibm.ws.tx.embeddable;version=latest,\
   com.ibm.websphere.javaee.annotation.1.2;version=latest,\
   com.ibm.wsspi.org.osgi.service.component.annotations
