+++ dev/com.ibm.ws.wsat.concurrent_fat/build.gradle
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2021, 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,5 +8,11 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
+
+// Define G:A:V coordinates of each dependency
+dependencies {
+  requiredLibs project(":com.ibm.ws.transaction.fat.util"),
+               project(':io.openliberty.org.apache.commons.logging')
+}
 
 addRequiredLibraries.dependsOn addJakartaTransformer
